+++ modules/aws/testcontainers/aws/__init__.py
@@ -1,0 +1,1 @@
+from .aws_lambda import AWSLambdaContainer  # noqa: F401