+++ modules/generic/tests/test_generic.py
@@ -7,17 +7,17 @@ from httpx import get
 
 from testcontainers.core.waiting_utils import wait_for_logs
 from testcontainers.core.image import DockerImage
-from testcontainers.core.generic import ServerContainer
+from testcontainers.generic import ServerContainer
 
 TEST_DIR = Path(__file__).parent
 
 
 @pytest.mark.parametrize("test_image_cleanup", [True, False])
 @pytest.mark.parametrize("test_image_tag", [None, "custom-image:test"])
-def test_srv_container(test_image_tag: Optional[str], test_image_cleanup: bool, check_for_image, port=9000):
+def test_server_container(test_image_tag: Optional[str], test_image_cleanup: bool, check_for_image, port=9000):
     with (
         DockerImage(
-            path=TEST_DIR / "image_fixtures/python_server",
+            path=TEST_DIR / "samples/python_server",
             tag=test_image_tag,
             clean_up=test_image_cleanup,
             #
@@ -37,8 +37,14 @@ def test_srv_container(test_image_tag: Optional[str], test_image_cleanup: bool, 
     check_for_image(image_short_id, test_image_cleanup)
 
 
+def test_server_container_no_port():
+    with pytest.raises(TypeError):
+        with ServerContainer(path="./modules/generic/tests/samples/python_server", tag="test-srv:latest"):
+            pass
+
+
 def test_like_doctest():
-    with DockerImage(path=TEST_DIR / "image_fixtures/python_server", tag="test-srv:latest") as image:
+    with DockerImage(path=TEST_DIR / "samples/python_server", tag="test-srv:latest") as image:
         with ServerContainer(port=9000, image=image) as srv:
             url = srv._create_connection_url()
             response = get(f"{url}", timeout=5)