+++ modules/rabbitmq/tests/test_rabbitmq.py
@@ -13,14 +13,17 @@ MESSAGE = {"hello": "world"}
 
 
 @pytest.mark.parametrize(
-    argnames=["port", "username", "password"],
+    argnames=["port", "username", "password", "vhost"],
     argvalues=[
-        [None, None, None],  # use the defaults
-        [5673, None, None],  # test with custom port
-        [None, "my_test_user", "my_secret_password"],  # test with custom credentials
+        [None, None, None, None],  # use the defaults
+        [5673, None, None, None],  # test with custom port
+        [None, "my_test_user", "my_secret_password", None],  # test with custom credentials
+        [None, None, None, "vhost"],  # test with custom vhost
     ],
 )
-def test_docker_run_rabbitmq(port: Optional[int], username: Optional[str], password: Optional[str]):
+def test_docker_run_rabbitmq(
+    port: Optional[int], username: Optional[str], password: Optional[str], vhost: Optional[str]
+):
     """Run rabbitmq test container and use it to deliver a simple message."""
     kwargs = {}
     if port is not None:
@@ -29,6 +32,8 @@ def test_docker_run_rabbitmq(port: Optional[int], username: Optional[str], passw
         kwargs["username"] = username
     if password is not None:
         kwargs["password"] = password
+    if vhost is not None:
+        kwargs["vhost"] = vhost
 
     rabbitmq_container = RabbitMqContainer("rabbitmq:latest", **kwargs)
     with rabbitmq_container as rabbitmq: