+++ modules/test_module_import/testcontainers/test_module_import/__init__.py
@@ -1,0 +1,1 @@
+from .new_sub_module import NewSubModuleContainer  # noqa: F401